moved tokencheck to handlers, since this is not a model
